Add skeleton fallbacks and suspense coverage for key flows
