+++ vitest.config.ts
@@ -4,12 +4,15 @@ import path from "path"
 export default defineConfig({
   test: {
     environment: "node",
-    include: ["tests/**/*.test.ts"],
+    include: ["tests/**/*.test.ts", "tests/**/*.test.tsx"],
   },
   resolve: {
     alias: {
       "@": path.resolve(__dirname),
       "server-only": path.resolve(__dirname, "tests/mocks/server-only.ts"),
     },
+  },
+  esbuild: {
+    jsx: "automatic",
   },
 })
